Merged PR 19: [backend] Get stripe customer ID on demand
## Let's say you have this

```
query {
  user(id: 2) {
    email
    alias
  }
}
```

We don't care if there is a `stripeCustomerId` yet. There's no need to create it when the user is created (the user might not ever spend $ or add a payment, meaning the customer ID isn’t required).

## BUT, what if we request `stripeCustomerId` in the graphql query?

```
query {
  user(id: 2) {
    email               // return field value `email` -> @Field()
    alias               // return field value `alias` -> @Field()
    stripeCustomerId    // custom @ResolveField()`
  }
}
```

What we want to expose is a function on `User`... to **"lazily" create the stripe customer ID only when requested the 1st time**. The query for `stripeCustomerId` actually invokes a custom `@ResolveField`  definition, `UserResolver.resolveStripeCustomerIdAsync()`

This in turn calls `usersService.getOrCreateCustomerId()`

This same function can be used if we want to get the stripe customer ID in code.

1. get a User from the database
2. call `UserService.getOrCreateCustomerId(user)`

Which does the same thing as if stripeCustomerId was requested from the GQL query.

Related work items: #266